feat: rename project to LF Tonegen and refactor architecture
* Rename project from "Tone Generator" to "LF Tonegen" across all files, resources, and UI titles.
* Refactor codebase to follow MVVM architecture with dedicated packages for audio, data, ui, and viewmodel.
* Implement SettingsRepository to isolate DataStore persistence logic.
* Replace deprecated ClipboardManager with modern LocalClipboard API and async ClipEntry handling.
* Fix frequency wheel bug by ensuring selected frequency snaps to valid range on limit changes.
* Enhance Settings UI with numeric-only input, clamping (10-30000 Hz), and focus-loss validation.
* Convert frequency range storage and processing from Float to Int for better precision and UI handling.
* Increase size of mic level meter, text, and save icon by 50% for improved visibility.
* Update project documentation (README.md, DEVELOPMENT.md) and establish English-only development standards.
* Increment version to 1.1.2 (versionCode 6) in build.gradle.kts

+++ app/src/main/java/dev/shard9/tonegenerator/ui/screens/GeneratorScreen.kt
@@ -132,7 +132,7 @@ fun GeneratorScreen(toneGenerator: ToneGenerator, viewModel: AppViewModel, modif
                         viewModel.updateSelectedFrequency(it)
                         toneGenerator.setFrequency(it.toDouble())
                     },
-                    range = viewModel.minFreq..viewModel.maxFreq,
+                    range = viewModel.minFreq.toFloat()..viewModel.maxFreq.toFloat(),
                     size = 240.dp,
                 )
